Update 01_00_ARC01_1_2.txt
I mentioned it in the file, but this one should probably be cross checked with 01_01_CIEL01_1_3.txt. Most of the text here is also in that one, but the translation differs in a few places. It's pretty minor and could be left as-is, since you'd really only notice if you played both scenes back to back.